Updated classes to use uint16_t instead of int, and std::string instead of const char*. Added helper missing functions for some enum classes

+++ FattyMenu-Lib/Source/Core/DistributionPermit/Package.cpp
@@ -2,7 +2,7 @@
 
 
 namespace FattyMenu {
-	CPackage::CPackage(const std::string& a_name, int a_cost, int a_quantity)
+	CPackage::CPackage(const std::string& a_name, uint16_t a_cost, uint16_t a_quantity)
 		: m_name(a_name), m_cost(a_cost), m_quantity(a_quantity) {
 	}
